backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic>=2.9.2
pydantic-settings>=2.4.0
firebase-admin==6.5.0
langchain==0.3.7
langchain-groq==0.2.3
langchain-community==0.3.7
anyio>=4.8.0
httpx>=0.28.1
python-multipart==0.0.9
